Trello 221 - updates to the data model following UI implementation, wireframes review with Fuel Suppliers
