+++ APISpec/gen/admin.py
@@ -25,6 +25,7 @@ from .models.CreditTrade import CreditTrade
 from .models.CreditTradeHistory import CreditTradeHistory
 from .models.CreditTradeStatus import CreditTradeStatus
 from .models.CreditTradeType import CreditTradeType
+from .models.CreditTradeZeroReason import CreditTradeZeroReason
 
 from .models.FuelSupplier import FuelSupplier
 from .models.FuelSupplierActionsType import FuelSupplierActionsType
@@ -36,14 +37,10 @@ from .models.FuelSupplierContact import FuelSupplierContact
 from .models.FuelSupplierContactRole import FuelSupplierContactRole
 from .models.FuelSupplierHistory import FuelSupplierHistory
 from .models.FuelSupplierStatus import FuelSupplierStatus
-from .models.FuelSupplierType import FuelSupplierType
 from .models.Notification import Notification
 from .models.NotificationEvent import NotificationEvent
 from .models.NotificationType import NotificationType
 
-from .models.Opportunity import Opportunity
-from .models.OpportunityHistory import OpportunityHistory
-from .models.OpportunityStatus import OpportunityStatus
 from .models.Permission import Permission
 
 from .models.Role import Role
@@ -65,6 +62,7 @@ admin.site.register(CreditTrade)
 admin.site.register(CreditTradeHistory)
 admin.site.register(CreditTradeStatus)
 admin.site.register(CreditTradeType)
+admin.site.register(CreditTradeZeroReason)
 admin.site.register(FuelSupplier)
 admin.site.register(FuelSupplierActionsType)
 admin.site.register(FuelSupplierAttachment)
@@ -75,13 +73,9 @@ admin.site.register(FuelSupplierContact)
 admin.site.register(FuelSupplierContactRole)
 admin.site.register(FuelSupplierHistory)
 admin.site.register(FuelSupplierStatus)
-admin.site.register(FuelSupplierType)
 admin.site.register(Notification)
 admin.site.register(NotificationEvent)
 admin.site.register(NotificationType)
-admin.site.register(Opportunity)
-admin.site.register(OpportunityHistory)
-admin.site.register(OpportunityStatus)
 admin.site.register(Permission)
 admin.site.register(Role)
 admin.site.register(RolePermission)
