--- APISpec/gen/models/OpportunityHistory.py
@@ -1,45 +1,0 @@
-"""
-    REST API Documentation for the NRS TFRS Credit Trading Application
-
-    The Transportation Fuels Reporting System is being designed to streamline compliance reporting for transportation fuel suppliers in accordance with the Renewable & Low Carbon Fuel Requirements Regulation.
-
-    OpenAPI spec version: v1
-        
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-"""
-
-import datetime
-
-from django.db import models
-from django.utils import timezone
-from .Opportunity import Opportunity
-from .User import User
-from .CreditTradeType import CreditTradeType
-from .FuelSupplierType import FuelSupplierType
-
-from auditable.models import Auditable
-
-class OpportunityHistory(Auditable):	    
-    opportunityFK = models.ForeignKey('Opportunity', related_name='OpportunityHistoryopportunityFK')   
-    userFK = models.ForeignKey('User', related_name='OpportunityHistoryuserFK')   
-    updateTime = models.DateField(blank=True, null=True)   
-    creditTradeTypeFK = models.ForeignKey('CreditTradeType', related_name='OpportunityHistorycreditTradeTypeFK')   
-    newNumberOfCredits = models.IntegerField()   
-    newProposedValuePerCredit = models.CharField(blank=True, null=True, max_length=255)   
-    fuelSupplierTypeFK = models.ForeignKey('FuelSupplierType', related_name='OpportunityHistoryfuelSupplierTypeFK')   
-    newTradeEffectiveDate = models.DateField(blank=True, null=True)   
-    note = models.CharField(max_length=4000, blank=True, null=True)   
-    class Meta:
-        db_table = 'OPPORTUNITY_HISTORY'
-
